v0.000000000000002

+++ Android/iTrash/app/src/main/java/com/hackupcteam/itrash/MainActivity.java
@@ -40,7 +40,7 @@ public class MainActivity extends AppCompatActivity {
     public ArrayList<String> ProductsAdded;
     private Button btn;
     private ListView lista1;
-    private int precio = 0;
+    private Double precio = 0.0;
 
 
     @Override
@@ -88,13 +88,13 @@ public class MainActivity extends AppCompatActivity {
                     int pos = ProductsAdded.indexOf(p.getName() + "\n" + p.getPrecio());
                     if (ProductsAdded.contains(p.getName() + "\n" + p.getPrecio()))
                         ProductsAdded.remove(pos);
-                        precio = precio - Integer.parseInt(p.getPrecio());
+                        precio = precio - Double.parseDouble(p.getPrecio());
                 } else {
                     System.out.print("Primer cop");
                     c.setChecked(true);
                     if (!ProductsAdded.contains(p.getName() + "\n" + p.getPrecio()))
                         ProductsAdded.add(p.getName() + "\n" + p.getPrecio());
-                        precio = precio + Integer.parseInt(p.getPrecio());
+                        precio = precio + Double.parseDouble(p.getPrecio());
                 }
             }
         });
